v3
Falta un sistema de duracion de cartas

+++ CatTarot/Assets/Scripts/Enemigo.cs
@@ -2,8 +2,8 @@ using UnityEngine;
 
 public class Enemigo : MonoBehaviour
 {
-    [SerializeField] public float vidaMaxima;
-    [SerializeField] public float vidaActual;
+    [SerializeField] public int vidaMaxima;
+    [SerializeField] public int vidaActual;
     // Start is called once before the first execution of Update after the MonoBehaviour is created
     void Start()
     {
